Implement file watch request via file browser
The file select input only gives a file name,
the drag and drop interface has a webkit extension that may give a path
but experimentally it was the basename prepended with /
so not useful for passing to notify.

Instead this implements a basic file browser
by making directory enumeration requests via ref websocket.