Rewrite to serve over HTTP rather than websocket
WebSockets do not participate in HTTP auth
and we can't configure browsers to use client certificates
to combine authentication and encryption.

Implementing authentication on top of web sockets
is a task I don't want to be responsible for,
so switching to an HTTP interface is a better idea.

SSE is an appropriate mechanism for sending file change notifications,
HTTP2 removes the downside of making extra connections,
and assocating responses with requests is an added benefit.

Unfortunately this loses the ability to have the server shut down
whent the cilent has disconnected.
This is an issue with async-h1, which just copies response data
into the connection without polling writability
and detecting that the writer has disconnected.
Fixing it to do otherwise has the hurdle of async-std apparently not
having a mechanism to do so,
which means that a fix probably requires using Tokio.

Hyper has a mechanism for clean shutdown,
which implies a level of tracking that could permit shutting down
after a given connection has ended,
but it has yet to be discerned whether connection ending is driven
from the client having disconnected or the server finishing serving.