agent: Send resolved path with read response
To have change notifications update the file
they need to be identified as the same.
The logical thing to use is file path,
but that means they have to be indexed by absolute path,
and the most reliable way to have read response do that
is have the response include the resolved path.